Initial project skeleton with a couple exercises complete
This is a simple c# project that I'll be using as I try the
exercises. Rather than do them in Java, I'll be using c# as it's
what I use at work and what I want to be the most familiar with. This
is all setup to be used on linux with mono so ymmv on windows.