Updated Graph adt to not be broke and basic dfs implementation.
Graph wasn't operating correctly as it didn't properly initialize
the adjacency list. Updating the entry point to provide the number
of nodes in the graph which is used for initialization. Also added
a basic dfs that provides Marked (i.e. IsConnected) interface. I'll
expand on that with a edgeTo[] to find paths next and then test for
coverage on that's there.