Dfs HasPath and PathTo Implementations
This adds a few basic Apis to make the dfs class more useful. Utilizing
the array initialization to answer the HasPath question, but you could
also simply use the marked array. I didn't like the idea of the edgeto
defaulting to zero because that doesn't mean there is a connection
to the root. Also had to do some hackery with the list.Reverse because
mono doesn't offer a stack implementation that is generic. Simple
tests included to verify basic operation.