Initial steps towards a OO parkinglot
This is to solve the design an OO parking lot problem. I wanted to try
and code it out to see how my initial modeling would work for an
implementation. I'm already finding cases where perhaps IVehicle
should be abstract because it simply is listing state like constraints.
Also, I am not sure how I'll organize the different types and map
them to eachother. For example, does the level keep track of the
different types of parking spots or do I had an enum / method on the
parking spot to see if it supports the given vehicle type. Leaning
more towards the former but hoping to flush that out.